Add servant prometheus metrics and remove wai
