+++ services/cargohold/test/integration/Main.hs
@@ -22,9 +22,6 @@ where
 
 import qualified API.V3
 import Bilge hiding (body, header)
-import qualified CargoHold.API (sitemap)
-import Data.Metrics.Test (pathsConsistencyCheck)
-import Data.Metrics.WaiRoute (treeToPaths)
 import Data.Proxy
 import Data.Tagged
 import Data.Text.Encoding (encodeUtf8)
@@ -33,10 +30,8 @@ import Imports hiding (local)
 import qualified Metrics
 import Network.HTTP.Client (responseTimeoutMicro)
 import Network.HTTP.Client.TLS
-import Network.Wai.Utilities.Server (compile)
 import Options.Applicative
 import Test.Tasty
-import Test.Tasty.HUnit
 import Test.Tasty.Options
 import TestSetup
 import Util.Options
@@ -86,12 +81,7 @@ main = runTests go
     go c i = withResource (getOpts c i) releaseOpts $ \opts ->
       testGroup
         "Cargohold"
-        [ testCase "sitemap" $
-            assertEqual
-              "inconcistent sitemap"
-              mempty
-              (pathsConsistencyCheck . treeToPaths . compile $ CargoHold.API.sitemap),
-          API.V3.tests opts,
+        [ API.V3.tests opts,
           Metrics.tests opts
         ]
     getOpts _ i = do
